split out job calculation logic into discreet funtions (tests still need to be moved)
