+++ utils/prepare_locations_utils/job_calculator/job_calculator.py
@@ -1,13 +1,19 @@
 import pyspark.sql.functions as F
 from pyspark.sql.types import IntegerType
 
-from utils.prepare_locations_utils.job_calculator.job_estimate_calculator import \
-    calculate_jobcount_totalstaff_equal_wkrrecs, \
-    calculate_jobcount_coalesce_totalstaff_wkrrecs, calculate_jobcount_abs_difference_within_range, \
-    calculate_jobcount_handle_tiny_values, calculate_jobcount_estimate_from_beds
+from utils.prepare_locations_utils.job_calculator.calculate_job_count_for_tiny_values import \
+    calculate_jobcount_handle_tiny_values
+from utils.prepare_locations_utils.job_calculator.calculate_jobcount_abs_difference_within_range import \
+    calculate_jobcount_abs_difference_within_range
+from utils.prepare_locations_utils.job_calculator.calculate_jobcount_coalesce_totalstaff_wkrrecs import \
+    calculate_jobcount_coalesce_totalstaff_wkrrecs
+from utils.prepare_locations_utils.job_calculator.calculate_jobcount_estimate_from_beds import \
+    calculate_jobcount_estimate_from_beds
+from utils.prepare_locations_utils.job_calculator.calculate_jobcount_total_staff_equal_worker_records import \
+    calculate_jobcount_totalstaff_equal_wkrrecs
 
 
-# TODO: write test cases for this so we are testing how this logic works together
+
 def calculate_jobcount(input_df):
     print("Calculating job_count...")
